Use default value '0' consistently for handshaking signals
Makes no difference in these cases, since the signals are assigned combinatorially anyway.
But we try to be consistent.

A default value of '1' on 'ready' would mean that a faulty implementation would pop inputs endlessly.
It would be preferrable if it would stall the upstream data, which would be easier to detect when debugging.

For 'valid' it would mean that transactions are sent endlessly back-to-back potentialy with garbage data that could lead to fatal side effects.

Note that in some places the default value has to be zero,
e.g. unsigned_divider.